add chart into a Word document

add a ms_chart into an rdocx object, the graphic will be inserted in an empty paragraph.

body_add_chart(x, chart, style = NULL, pos = "after", width = 5,
  height = 3)

Arguments

x
an rdocx object

chart
an ms_chart object.

style
paragraph style

pos
where to add the new element relative to the cursor, one of "after", "before", "on".

height, width
height and width in inches.

Examples

library(officer)
my_barchart <- ms_barchart(data = browser_data,
  x = "browser", y = "value", group = "serie")
my_barchart <- chart_settings( my_barchart, grouping = "stacked",
  gap_width = 50, overlap = 100 )

doc <- read_docx()
doc <- body_add_chart(doc, chart = my_barchart, style = "centered")
print(doc, target = "barchart_example.docx")

#> [1] "/Users/davidgohel/gitprojects/mschart/docs/reference/barchart_example.docx"