Smooth series

Specify mappings from levels in the data to smooth or not lines. This feature only applies to ms_linechart().

chart_data_smooth(x, values)

Arguments

x

an ms_chart object.

values

integer(num of series): a set of smooth values to map data values to. It is a named vector, the values will be matched based on the names. Possible values are 0 or 1 If it contains only one integer it will be associated to all existing series.

See also

Examples

linec <- ms_linechart(data = iris, x = "Sepal.Length",
  y = "Sepal.Width", group = "Species")
linec <- chart_data_smooth(linec,
  values = c(virginica = 0, versicolor = 0, setosa = 0) )